+++ packages/widgetbook/test/generator/custom_args/custom_args.stories.g.dart
@@ -29,6 +29,7 @@ class LabelBadgeStory extends Story<LabelBadge, NumericBadgeInputArgs> {
     NumericBadgeInputArgs? args,
     StoryWidgetBuilder<LabelBadge, NumericBadgeInputArgs>? builder,
     super.scenarios,
+    super.excludeFromTests,
   }) : super(
          args: args ?? NumericBadgeInputArgs(),
          builder: builder ?? defaults.builder!,