+++ packages/widgetbook/test/generator/generic_bound_dynamic/generic_bound_dynamic.stories.g.dart
@@ -36,6 +36,7 @@ class DynamicBoundWidgetStory<D, T extends Map<dynamic, D>>
     StoryWidgetBuilder<DynamicBoundWidget<D, T>, DynamicBoundWidgetArgs<D, T>>?
     builder,
     super.scenarios,
+    super.excludeFromTests,
   }) : super(
          builder:
              builder ??