+++ packages/widgetbook/test/generator/static_methods/static_methods.stories.g.dart
@@ -34,6 +34,7 @@ class StaticMethodWidgetStory
     StaticMethodWidgetArgs? args,
     StoryWidgetBuilder<StaticMethodWidget, StaticMethodWidgetArgs>? builder,
     super.scenarios,
+    super.excludeFromTests,
   }) : super(
          args: args ?? StaticMethodWidgetArgs(),
          builder: